models/ECSpell/requirements.txt
botocore==1.20.103
boto3==1.17.103
paddlepaddle_gpu==2.1.2
jieba==0.42.1
PyTrie==0.4.0
datasets==1.12.1
pandas==1.2.4
requests==2.25.1
pywubi==0.0.2
zhconv==1.4.1
torchvision==0.10.0
pypinyin==0.42.0
transformers==4.11.0
packaging==21.0
tokenizers==0.10.3
tqdm==4.62.3
numpy==1.20.1
overrides==6.1.0
conllu==4.4
h5py==3.3.0
allennlp==2.9.2
apex==0.9.10dev
fairscale==0.4.6
opencc==1.1.3
optuna==2.10.0
paddle==1.0.2
Pillow==9.1.0
ray==1.11.0
scikit_learn==1.0.2
tensorboardX==2.5
torchtext==0.12.0

botocore==1.20.103
boto3==1.17.103
paddlepaddle_gpu==2.1.2
jieba==0.42.1
PyTrie==0.4.0
datasets==1.12.1
pandas==1.2.4
requests==2.25.1
pywubi==0.0.2
zhconv==1.4.1
torchvision==0.10.0
pypinyin==0.42.0
transformers==4.11.0
packaging==21.0
tokenizers==0.10.3
tqdm==4.62.3
numpy==1.20.1
overrides==6.1.0
conllu==4.4
h5py==3.3.0
allennlp==2.9.2
apex==0.9.10dev
fairscale==0.4.6
opencc==1.1.3
optuna==2.10.0
paddle==1.0.2
Pillow==9.1.0
ray==1.11.0
scikit_learn==1.0.2
tensorboardX==2.5
torchtext==0.12.0

